nvmf: fix set-features host-behavior-support -validate reserved field
Motivation:
See in NVM Express: "Receipt of reserved coded values in defined fields
in commands shall be reported as an error"
The real issue with code before fix was that ESX initator that supports
XCOPY sent Set-feature HBS to mode that supports XCOPY. Target accepted
the command since 'struct spdk_nvme_host_behavior' was extended in TP
supporting XCOPY. As a result initiator tries to perform hardware
accelerated xcopy that is not supported by the target. After applying
the fix ESX initiator for volume copy instead of using modern XCOPY
performs bunch of reads and writes.

Change-Id: I9552e7b2bdc214c3309a1d51549b2e4398f62f04
Signed-off-by: Leonid Chernin <leonidc@il.ibm.com>